[docs] 영상 &  자막 연동 진행 중
