--- src/components/study/StudySubtitles.jsx
@@ -1,5 +1,0 @@
-function StudySubtitles() {
-  return <div>StudySubtitles</div>;
-}
-
-export default StudySubtitles;
